+++ frontend/ui/package.json
@@ -45,7 +45,7 @@
     "./widgets": "./src/widgets/index.ts",
     "./date-picker": "./src/components/date-picker.tsx",
     "./onboarding": "./src/components/onboarding/index.tsx",
-    "./tanstak-table": "./src/components/tanstak-table/index.ts"
+    "./tanstack-table": "./src/components/tanstack-table/index.ts"
   },
   "dependencies": {
     "@atlaskit/pragmatic-drag-and-drop": "1.4.0",
