+++ frontend/ui/src/components/tanstack-table/index.ts
@@ -2,4 +2,4 @@
 export { default as UsersTableDemo } from "./users-demo";
 export { default as TasksTableDemo } from "./tasks-demo";
 export { default as SidebarFilterDemo } from "./sidebarfilter-demo";
-export { DataTable } from "./tanstak-table";
+export { DataTable } from "./tanstack-table";
