+++ frontend/ui/src/components/tanstack-table/tasks-demo.tsx
@@ -1,12 +1,11 @@
 "use client"
 
 import React, { useState, useEffect, useCallback } from "react"
-import { DataTable } from "./tanstak-table"
+import { DataTable } from "./tanstack-table"
 import { registerCellRenderer } from "./cell-renderers"
-import { RowAction, SidebarFilterConfig } from "./types"
-import { Calendar, Clock, FileText, UserCircle, Tag, BarChart4, Check } from "lucide-react"
+import { RowAction } from "./types"
 
-// Custom rating cell renderer
+// Custom rating cell renderer (example of extending the table with a new column type)
 const RatingCell: React.FC<{ value: number }> = ({ value }) => {
   return (
     <div className="flex items-center">
@@ -150,6 +149,7 @@ const USER_TABLE_COLUMNS = [
     accessorKey: "joinDate" as const,
     id: "joinDate",
     enableSorting: true,
+    // Example of custom date formatting
     format: {
       dateFormat: "YYYY-MM-DD HH:mm"
     }
@@ -179,6 +179,7 @@ const USER_TABLE_COLUMNS = [
     accessorKey: "balance" as const,
     id: "balance",
     enableSorting: true,
+    // Example of custom currency formatting
     format: {
       numberFormat: {
         style: "currency",
@@ -197,8 +198,16 @@ const USER_TABLE_COLUMNS = [
   }
 ]
 
-// Faceted filter definitions (for top filters)
+// Filter definitions
 const USER_TABLE_FACETS = [
+  {
+    column: "isActive",
+    title: "Status",
+    options: [
+      { label: "Active", value: true },
+      { label: "Inactive", value: false }
+    ]
+  },
   {
     column: "plan",
     title: "Plan",
@@ -206,6 +215,15 @@ const USER_TABLE_FACETS = [
       { label: "Success", value: "success" },
       { label: "Pending", value: "pending" },
       { label: "Failed", value: "failed" }
+    ]
+  },
+  {
+    column: "tags",
+    title: "Role",
+    options: [
+      { label: "Admin", value: "admin" },
+      { label: "Developer", value: "developer" },
+      { label: "Customer", value: "customer" }
     ]
   },
   {
@@ -221,59 +239,10 @@ const USER_TABLE_FACETS = [
   }
 ]
 
-// Sidebar filter definitions
-const SIDEBAR_FILTERS: SidebarFilterConfig<User>[] = [
-  {
-    type: "dateRange",
-    column: "joinDate",
-    title: "Time Range",
-    icon: <Calendar className="h-4 w-4" />,
-    initialCollapsed: false
-  },
-  {
-    type: "text",
-    column: "name",
-    title: "Name",
-    icon: <UserCircle className="h-4 w-4" />,
-    initialCollapsed: false
-  },
-  {
-    type: "boolean",
-    column: "isActive",
-    title: "Active",
-    icon: <Check className="h-4 w-4" />,
-    initialCollapsed: false
-  },
-  {
-    type: "multiSelect",
-    column: "tags",
-    title: "Tags",
-    icon: <Tag className="h-4 w-4" />,
-    options: [
-      { label: "Admin", value: "admin" },
-      { label: "Developer", value: "developer" },
-      { label: "Customer", value: "customer" }
-    ],
-    initialCollapsed: false
-  },
-  {
-    type: "select",
-    column: "plan",
-    title: "Plan",
-    icon: <FileText className="h-4 w-4" />,
-    options: [
-      { label: "Success", value: "success" },
-      { label: "Pending", value: "pending" },
-      { label: "Failed", value: "failed" }
-    ],
-    initialCollapsed: false
-  }
-]
-
 /**
- * Enhanced UsersTableDemo with Sidebar Filters
+ * Enhanced UsersTableDemo - Example usage of the improved DataTable component
  */
-const SidebarFilterDemo = () => {
+const EnhancedUsersTableDemo = () => {
   // Pagination state
   const [currentPage, setCurrentPage] = useState(0)
   const [pageSize, setPageSize] = useState(10)
@@ -408,14 +377,7 @@ const SidebarFilterDemo = () => {
         filterColumn="name"
         filterPlaceholder="Filter by name..."
         rowActions={getRowActions}
-
-        // Top faceted filters
         facets={USER_TABLE_FACETS}
-
-        // NEW: Sidebar filters
-        enableSidebarFilters={true}
-        sidebarFilters={SIDEBAR_FILTERS}
-        initialSidebarOpen={true}
 
         // Server pagination props
         serverPagination={true}
@@ -447,4 +409,4 @@ const SidebarFilterDemo = () => {
   )
 }
 
-export default SidebarFilterDemo
+export default EnhancedUsersTableDemo
